Simple Electricity Bill
Alice received her electricity bill for the month of October. The previous electricity units
of Alice bill were X units and the present units were Y units. The X or previous units is
the cumulative units for the month September and all previous months. Similarly, Y or
present units is the cumulative units consumed in October plus all previous months
(September, August, ……...). Calculate the units consumed in the month of October and
also calculate the total bill. Your program should print the units consumed in October
and the total bill for October.
1 unit price= 5 Rs.