Simple Restaurant Bill
Write a program that calculates the tax and tip on a restaurant bill for a customer with
a X$ meal charge.  The tax should be 4.75% of the meal cost.  The tip should be 10% of
the total after adding the tax.  Display the meal cost, tax amount, tip amount, and
total bill on the screen.