Distance between two cyclists.
Ali and Haider are two cyclists. Ali covers a distance of X km in Y hours and Z minutes.
While Haider covers a distance of A km in B hours and C minutes. Write a program
that calculates the speed of two cyclists and prints the speed of both players in m/sec.